+++ appsettings.json
@@ -19,7 +19,7 @@
     "ShowRenewButton": false,
     "NamespaceConstraints": null,
     "InitialSleepOnFailure": 30,
-    "PreferredChallengeType": "http"
+    "ChallengeType": "http"
   },
   "Acme": {
     "DirUrl": "https://acme-staging-v02.api.letsencrypt.org/directory",
